+++ project 3/src/ShortestPathOpdracht/Stap.java
@@ -1,6 +1,6 @@
 package ShortestPathOpdracht;
 
-abstract class Stap {
+abstract class Stap { // Deze class is bestaan alleen maar als tussenverbinding tussen de super en de sub classen.
     public abstract node returnNode();
-    public abstract int returnAfstand();
+    public abstract double returnAfstand();
 }
